requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.4
django-haystack==3.2.1
elastic-transport==8.1.2
elasticsearch==7.17.3
h11==0.13.0
idna==3.3
lxml==4.8.0
mysqlclient==2.1.0
numpy==1.22.3
outcome==1.1.0
pandas==1.4.2
pycparser==2.21
pygame==2.1.2
pymongo==4.1.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
selenium==4.1.3
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
